+++ js/main.js
@@ -3,7 +3,8 @@ import {renderUsersMiniatures} from './users_photo_miniatures.js';
 import './users_photo_big.js';
 import './comments.js';
 import './util.js';
-import './form.js';
+import './form_open.js';
+import './form_validation.js';
 
 
 // Задаём переменную-хранилище для нашего готового массива с фотографиями
